pyproject.toml
[project]
name = "web-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleh Matiukha",email = "olehmatiukha@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker (>=37.0.0,<38.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
